update readme. rename file
