+++ torchao/prototype/quantized_training/__init__.py
@@ -1,1 +1,1 @@
-from .subclass import Int8QTLinearWeight, int8_weight_only_quantized_training
+from .int8 import Int8QTLinearWeight, int8_weight_only_quantized_training
